+++ deno.jsonc
@@ -1,10 +1,6 @@
 {
   "compilerOptions": {
     "checkJs": false,
-    "jsx": "react",
-    "jsxFactory": "h",
-    "jsxFragmentFactory": "Fragment",
-    "lib": ["deno.ns", "dom"],
     "noFallthroughCasesInSwitch": true,
     "noImplicitOverride": true,
     "noImplicitReturns": true,
@@ -36,7 +32,7 @@
     "udd": "deno task star && deno run -A https://deno.land/x/udd@0.5.0/main.ts target/star.ts",
     "dnt": "deno run -A _tasks/dnt.ts",
     "test": "deno task ctx -- deno test -A -L=info --ignore=target --parallel -r=http://localhost:4646/",
-    "text:examples": "deno task test _tasks/test_examples.ts", 
+    "text:examples": "deno task test _tasks/test_examples.ts",
     "test:update": "deno task test -- --update",
     "moderate": "deno run -A https://deno.land/x/moderate@0.0.5/mod.ts && dprint fmt",
     "dev": "deno run -A --watch=deno.jsonc,docs/ main.ts --serve",
