--- deps/preact_render_to_string.ts
@@ -1,1 +1,0 @@
-export * from "https://esm.sh/preact-render-to-string@5.1.19?deps=preact@10.5.15"
